Move doobie syntax to a new trait
