+++ src/main/scala/thewho/repository/DoobieRepository.scala
@@ -1,8 +1,6 @@
 package thewho.repository
 
-import doobie._
-import doobie.util.Read
-import scalaz.zio.interop.catz._
+import doobie.util.transactor.{ Transactor => DoobieTransactor }
 import scalaz.zio.{ IO, Task }
 import thewho.auth.{ Credential, CredentialId, User, UserId }
 import thewho.repository.Repository.Service
@@ -10,11 +8,15 @@ import thewho.repository.error._
 
 trait DoobieRepository extends Repository {
 
-  protected def xa: Transactor[Task]
+  protected def xa: DoobieTransactor[Task]
+
+  object dsl extends syntax {
+    override protected def _xa: DoobieTransactor[Task] = xa
+  }
+
+  import dsl._
 
   override final val repository = new Service[Any] {
-
-    import syntax._
 
     override final def heartBeat: IO[RepositoryFailure, Unit] = sql.heartBeat._query[Unit].map(_ => ())
 
@@ -61,69 +63,6 @@ trait DoobieRepository extends Repository {
 
     override final def deleteCredential(credentialId: CredentialId): IO[RepositoryException, CredentialId] =
       sql.credential.delete(credentialId)._updateOrErrorWith(NoCredentialsUpdated).map(_ => credentialId)
-
-  }
-
-  object syntax {
-
-    import doobie.implicits._
-
-    type UpdatedRowsCount = Int
-
-    implicit class FragmentSyntax(fragment: Fragment) {
-
-      /**
-       * Run the sql query succeeding with an Option.
-       *
-       * @tparam A The type of the value returned inside the Option
-       */
-      def _query[A: Read]: IO[RepositoryFailure, Option[A]] =
-        fragment.query[A].option.transact(xa).mapError(CommonRepositoryFailure)
-
-      /**
-       * The same as [[_query]] but with a default value in the error channel or
-       * the unwrapped result in the success channel.
-       *
-       * @param ifEmpty The default value for th error channel
-       * @tparam A The type of the value returned in the success channel
-       */
-      def _queryOrErrorWith[A: Read](ifEmpty: RepositoryError): IO[RepositoryException, A] =
-        _query[A] >>= fromOption(ifEmpty)
-
-      /**
-       * Run the sql query returning the number of rows affected.
-       */
-      def _update: IO[RepositoryFailure, UpdatedRowsCount] =
-        fragment.update.run.transact(xa).mapError(CommonRepositoryFailure)
-
-      /**
-       * The same as [[_update]] but failing with the provided error if no rows were affected.
-       *
-       * @param ifNoHits The error in case of no rows affected
-       */
-      def _updateOrErrorWith(ifNoHits: RepositoryError): IO[RepositoryException, UpdatedRowsCount] =
-        _update >>= orErrorWith(ifNoHits)
-
-      /**
-       * Run the sql query returning the key of the value generated by the database.
-       *
-       * @tparam A The type of the generated key
-       */
-      def _create[A: Read]: IO[RepositoryFailure, A] =
-        fragment.update.withUniqueGeneratedKeys[A]("id").transact(xa).mapError(CommonRepositoryFailure)
-
-    }
-
-    private def fromOption[A](ifEmpty: RepositoryError)(option: Option[A]): IO[RepositoryError, A] =
-      option.fold[IO[RepositoryError, A]](IO fail ifEmpty)(IO succeed)
-
-    private def orErrorWith(
-      ifError: RepositoryError
-    )(updatedRowsCount: UpdatedRowsCount): IO[RepositoryError, UpdatedRowsCount] =
-      updatedRowsCount match {
-        case 0 => Task fail ifError
-        case i => Task succeed i
-      }
 
   }
